pyproject.toml
[tool.poetry]
name = "FasterAPI"
version = "0.0.39"
description = ""
authors = ["ulfaric <ryf0510@live.com>"]
readme = "README.md"
urls = { github = "https://github.com/ulfaric/FasterAPI/issues", documentation = "https://ulfaric.github.io/FasterAPI/" }

[tool.poetry.dependencies]
python = "^3.8"
fastapi = { extras = ["all"], version = "^0.110.0" }
colorlog = "^6.8.2"
sqlalchemy = "^2.0.28"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
bcrypt = "4.0.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
akatosh = "^3.0.7"
opentelemetry-sdk = "^1.24.0"
opentelemetry-api = "^1.24.0"
opentelemetry-instrumentation = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
